add test

+++ src/app/components/wallets/wallet-detail/wallet-detail.component.ts
@@ -1,6 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from "@angular/router";
 import {WalletService} from "../../../services/wallets/wallet.service";
+import {ToastrService} from "ngx-toastr";
 
 @Component({
   selector: 'app-wallet-detail',
@@ -13,7 +14,8 @@ export class WalletDetailComponent implements OnInit {
 
   constructor(protected activatedRouter: ActivatedRoute,
               protected walletService: WalletService,
-              protected router: Router) {
+              protected router: Router,
+              protected toastr: ToastrService) {
   }
 
   ngOnInit(): void {
@@ -26,10 +28,15 @@ export class WalletDetailComponent implements OnInit {
       this.walletById = res.data;
     })
   }
+
   deleteWallet(walletId: any){
-    this.walletService.delete(walletId).subscribe((res:any)=>{
-      console.log(res)
-    })
+    if(confirm('Are you sure ?!')){
+      this.walletService.delete(walletId).subscribe((res:any)=>{
+        this.toastr.success('Delete successfully')
+        this.router.navigate(['wallet/info'])
+      })
+    }
+
   }
 
   back(){
